+++ core/language/it/gallery.php
@@ -142,6 +142,7 @@ $lang = array_merge($lang, [
 	'IMAGE_INSERTED'               => 'Immagine inserita',
 	'IMAGE_LOCKED'                 => 'Spiacente, quest’immagine è bloccata. Non si può più commentare quest’immagine.',
 	'IMAGE_NAME'                   => 'Nome immagine',
+	'IMAGE_UPLOAD_TIME'            => 'Data di caricamento',
 	'IMAGE_NOT_EXIST'              => 'Quest’immagine non esiste.',
 	'HOTLINK_NOT_ALLOWED'          => 'I collegamenti diretti alle immagini non sono consentiti.',
 	'IMAGE_NOT_APPROVED'           => 'Per approvazione',